requirements.txt
pandas == 2.2.1
Flask == 3.0.2
matplotlib == 3.8.4
plotly == 5.22.0
streamlit == 1.34.0 
python-dotenv
gspread == 6.1.2
gspread-dataframe
